# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dimension {
  /** @minimum 1 */
  id: number;
  /**
     * @maximum 6
     * @exclusiveMinimum 0
     */
  length: number;
  createdAt: string;
}

export interface DimensionInput {
  /**
     * @maximum 6
     * @exclusiveMinimum 0
     */
  length: number;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface AdminSessionInput {
  /** @minLength 1 */
  pin: string;
}

export interface CutRequest {
  /** @minimum 1 */
  dimensionId: number;
  /**
     * @maximum 6
     * @exclusiveMinimum 0
     */
  length: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CutPlanInput {
  /** @minItems 1 */
  requests: CutRequest[];
}

export interface CutPiece {
  /** @minimum 1 */
  dimensionId: number;
  /** @exclusiveMinimum 0 */
  length: number;
}

export interface RailPlan {
  /** @minimum 1 */
  railNumber: number;
  pieces: CutPiece[];
  /** @minimum 0 */
  used: number;
  /** @minimum 0 */
  waste: number;
  /** @minimum 0 */
  reusableOffcut: number;
}

export interface CutPlan {
  stockLength: 6;
  /** @minimum 0 */
  totalRequested: number;
  /** @minimum 0 */
  railCount: number;
  /** @minimum 0 */
  totalWaste: number;
  /** @minimum 0 */
  totalReusableOffcut: number;
  rails: RailPlan[];
}

export interface Error {
  error: string;
}