# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dimensions
    description: Reusable aluminium rail cut dimensions
  - name: cutting
    description: Low-waste rail cutting calculations
  - name: admin
    description: Admin-only dimension management session
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dimensions:
    get:
      operationId: listDimensions
      tags: [dimensions]
      summary: List saved cut dimensions
      responses:
        "200":
          description: Saved dimensions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dimension"
    post:
      operationId: createDimension
      tags: [dimensions]
      summary: Add a reusable cut dimension
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DimensionInput"
      responses:
        "201":
          description: Created dimension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dimension"
        "400":
          description: Invalid dimension
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dimensions/{id}:
    delete:
      operationId: deleteDimension
      tags: [dimensions]
      summary: Delete a saved cut dimension
      parameters:
        - $ref: "#/components/parameters/DimensionId"
      responses:
        "204":
          description: Dimension deleted
        "404":
          description: Dimension not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check whether the current browser has an admin session
      responses:
        "200":
          description: Current admin session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
    post:
      operationId: unlockAdminSession
      tags: [admin]
      summary: Unlock admin-only dimension management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSessionInput"
      responses:
        "200":
          description: Admin session unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Incorrect PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: lockAdminSession
      tags: [admin]
      summary: Lock the current admin session
      responses:
        "204":
          description: Admin session locked
  /cut-plans:
    post:
      operationId: calculateCutPlan
      tags: [cutting]
      summary: Calculate a low-waste cutting plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CutPlanInput"
      responses:
        "200":
          description: Optimized cutting plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CutPlan"
        "400":
          description: Invalid cutting request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    DimensionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Dimension:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        length:
          type: number
          exclusiveMinimum: 0
          maximum: 6
        createdAt:
          type: string
          format: date-time
      required: [id, length, createdAt]
    DimensionInput:
      type: object
      properties:
        length:
          type: number
          exclusiveMinimum: 0
          maximum: 6
      required: [length]
    AdminSession:
      type: object
      properties:
        authenticated:
          type: boolean
      required: [authenticated]
    AdminSessionInput:
      type: object
      properties:
        pin:
          type: string
          minLength: 1
      required: [pin]
    CutRequest:
      type: object
      properties:
        dimensionId:
          type: integer
          minimum: 1
        length:
          type: number
          exclusiveMinimum: 0
          maximum: 6
        quantity:
          type: integer
          minimum: 1
      required: [dimensionId, length, quantity]
    CutPlanInput:
      type: object
      properties:
        requests:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CutRequest"
      required: [requests]
    CutPiece:
      type: object
      properties:
        dimensionId:
          type: integer
          minimum: 1
        length:
          type: number
          exclusiveMinimum: 0
      required: [dimensionId, length]
    RailPlan:
      type: object
      properties:
        railNumber:
          type: integer
          minimum: 1
        pieces:
          type: array
          items:
            $ref: "#/components/schemas/CutPiece"
        used:
          type: number
          minimum: 0
        waste:
          type: number
          minimum: 0
        reusableOffcut:
          type: number
          minimum: 0
      required: [railNumber, pieces, used, waste, reusableOffcut]
    CutPlan:
      type: object
      properties:
        stockLength:
          type: number
          const: 6
        totalRequested:
          type: number
          minimum: 0
        railCount:
          type: integer
          minimum: 0
        totalWaste:
          type: number
          minimum: 0
        totalReusableOffcut:
          type: number
          minimum: 0
        rails:
          type: array
          items:
            $ref: "#/components/schemas/RailPlan"
      required: [stockLength, totalRequested, railCount, totalWaste, totalReusableOffcut, rails]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]